Extract async proxy strategies to separate file
