+++ rails_event_store/lib/rails_event_store/all.rb
@@ -1,4 +1,5 @@
 require 'ruby_event_store'
+require 'rails_event_store/async_proxy_strategy'
 require 'rails_event_store/async_dispatcher'
 require 'rails_event_store/active_job_dispatcher'
 require 'rails_event_store/async_handler_helpers'
